+++ spec/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.2].define(version: 2025_04_06_213935) do
+ActiveRecord::Schema[7.2].define(version: 2025_04_26_061729) do
   create_table "comments", force: :cascade do |t|
     t.text "content", null: false
     t.integer "user_id"
@@ -78,6 +78,18 @@ ActiveRecord::Schema[7.2].define(version: 2025_04_06_213935) do
     t.index ["actor_id"], name: "index_federails_followings_on_actor_id"
     t.index ["target_actor_id"], name: "index_federails_followings_on_target_actor_id"
     t.index ["uuid"], name: "index_federails_followings_on_uuid", unique: true
+  end
+
+  create_table "federails_hosts", force: :cascade do |t|
+    t.string "domain", null: false
+    t.string "nodeinfo_url"
+    t.string "software_name"
+    t.string "software_version"
+    t.text "protocols", default: "[]"
+    t.text "services", default: "{}"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["domain"], name: "index_federails_hosts_on_domain", unique: true
   end
 
   create_table "posts", force: :cascade do |t|